refactor(native): extract browser lifecycle state
